requirements.txt

discord.py>=2.0
requests>=2.28
python-dotenv>=1.0
pytest>=7.0
pytest-asyncio>=0.20
beautifulsoup4>=4.12
